+++ src/main/java/com/formssafe/domain/question/repository/ObjectiveQuestionRepository.java
@@ -3,5 +3,5 @@ package com.formssafe.domain.question.repository;
 import com.formssafe.domain.question.entity.ObjectiveQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObjectiveQuestionRepository extends JpaRepository<ObjectiveQuestion, Integer> {
+public interface ObjectiveQuestionRepository extends JpaRepository<ObjectiveQuestion, Long> {
 }
